ui: code gardening and DRYing up the exporter status component
