+++ packages/app/src/client/integrations/exporterCloudWatch/Form.tsx
@@ -165,3 +165,5 @@ export const ExporterCloudWatchForm = ({ handleCreate }: Props) => {
     </Box>
   );
 };
+
+export default ExporterCloudWatchForm;
