+++ packages/app/src/client/integrations/exporterCloudWatch/Show.tsx
@@ -50,11 +50,6 @@ export const ExporterCloudWatchShow = () => {
     ];
   }, [integration?.grafana_metadata]);
 
-  if (!integration) {
-    // TODO: add loading or NotFound here
-    return null;
-  }
-
   // NOTE: the timeranges for these urls is not the same as that used by the status component
   // const errorLogsUrl = useMemo(() => {
   //   const path = `orgId=1&left=%5B%22now-1h%22,%22now%22,%22logs%22,%7B%22expr%22:%22%7Bk8s_namespace_name%3D%5C%22${tenant.name}-tenant%5C%22,k8s_container_name%3D%5C%22exporter%5C%22,k8s_pod_name%3D~%5C%22%5Eexporter-${integration.key}-%5Ba-z0-9-%5D*%5C%22%7D%20%7C%3D%20%5C%22stderr%5C%22%20%7C%3D%20%5C%22${ERROR_STR}%5C%22%22%7D%5D`;
@@ -62,9 +57,16 @@ export const ExporterCloudWatchShow = () => {
   // }, [tenant.name, integration.key]);
 
   const logsUrl = useMemo(() => {
-    const path = `orgId=1&left=%5B%22now-1h%22,%22now%22,%22logs%22,%7B%22expr%22:%22%7Bk8s_namespace_name%3D%5C%22${tenant.name}-tenant%5C%22,k8s_container_name%3D%5C%22exporter%5C%22,k8s_pod_name%3D~%5C%22%5Eexporter-${integration.key}-%5Ba-z0-9-%5D*%5C%22%7D%22%7D%5D`;
-    return `${window.location.protocol}//system.${window.location.host}/grafana/explore?${path}`;
-  }, [tenant.name, integration.key]);
+    if (integration?.key) {
+      const path = `orgId=1&left=%5B%22now-1h%22,%22now%22,%22logs%22,%7B%22expr%22:%22%7Bk8s_namespace_name%3D%5C%22${tenant.name}-tenant%5C%22,k8s_container_name%3D%5C%22exporter%5C%22,k8s_pod_name%3D~%5C%22%5Eexporter-${integration.key}-%5Ba-z0-9-%5D*%5C%22%7D%22%7D%5D`;
+      return `${window.location.protocol}//system.${window.location.host}/grafana/explore?${path}`;
+    } else return "";
+  }, [tenant.name, integration?.key]);
+
+  if (!integration) {
+    // TODO: add loading or NotFound here
+    return null;
+  }
 
   const integrationDef = integrationDefRecords[integration.kind];
 
@@ -139,3 +141,5 @@ export const ExporterCloudWatchShow = () => {
     </>
   );
 };
+
+export default ExporterCloudWatchShow;
