+++ packages/app/src/client/integrations/exporterStackdriver/Show.tsx
@@ -121,3 +121,5 @@ export const ExporterStackdriverShow = () => {
     </>
   );
 };
+
+export default ExporterStackdriverShow;
